refactor: move `command` to `discord.command`
