+++ src/main/kotlin/com/github/frozensync/discord/command/CommandArgs.kt
@@ -1,4 +1,4 @@
-package com.github.frozensync.command
+package com.github.frozensync.discord.command
 
 import java.util.*
