+++ src/main/kotlin/com/github/frozensync/discord/command/CommandSet.kt
@@ -1,4 +1,4 @@
-package com.github.frozensync.command
+package com.github.frozensync.discord.command
 
 interface CommandSet {
     val commands: Map<String, Command>
