+++ src/main/kotlin/com/github/frozensync/monitoring/MonitoringCommandSet.kt
@@ -1,7 +1,7 @@
 package com.github.frozensync.monitoring
 
-import com.github.frozensync.command.Command
-import com.github.frozensync.command.CommandSet
+import com.github.frozensync.discord.command.Command
+import com.github.frozensync.discord.command.CommandSet
 import kotlinx.coroutines.reactive.awaitFirst
 import java.lang.management.ManagementFactory
 import java.time.Duration
